+++ tests/testthat/test-parse-block.R
@@ -8,6 +8,8 @@ test_that("trimws works", {
 
 test_that("plumbBlock works", {
   lines <- c(
+    "#* Plumber comments",
+    "# Normals comments",
     "#' @get /",
     "#' @post /",
     "#' @filter test",
@@ -17,6 +19,7 @@ test_that("plumbBlock works", {
   expect_equal(b$paths[[1]], list(verb="POST", path="/"))
   expect_equal(b$paths[[2]], list(verb="GET", path="/"))
   expect_equal(b$filter, "test")
+  expect_equal(b$comments, " Plumber comments")
 
   # due to covr changing some code, the return answer is very strange
   # the tests below should be skipped on covr