+++ NEWS.md
@@ -8,6 +8,7 @@ plumber 1.0.0.9999 Development version
 ### Bug fixes
 
 * When calling `Plumber$handle()` and defining a new `PlumberEndpoint`, `...` will be checked for invalid names #677
+* Ignore regular comments in block parsing
 
 * `/__swagger__/` now always redirect to `/__docs__/`, even when Swagger isn't the selected interface. Use `options(plumber.legacyRedirects = FALSE)` to disable this behavior (@blairj09 #694)